refactor: decouple api auth from services to satisfy layer guard
The 12B import switch repointed two api files at services.auth directly,
which the layer guard forbids (api -> di, core only; never services).
Resolve it properly rather than bridging through di:

- StateCookieSerializer / StateCookiePayload are pure (dataclasses +
  itsdangerous, no I/O), so move them to core/auth/state_cookie.py.
  api/routers/auth/oidc.py now imports from core (api -> core, allowed);
  services/auth re-exports them so existing services consumers are
  unaffected; the components.auth shim repoints to core.

- refresh_session_if_needed is infrastructure and stays in services, but
  the middleware reached it via a direct module import. Expose it as
  AuthService.refresh_session_if_needed (a thin seam over the helper,
  passing self) so the middleware calls it on the AuthService it already
  obtains from di — consistent with every other session op it performs.
  Update the middleware test: the mock auth-service gains a default
  refresh delegating to the real helper, and the two refresh-outcome
  tests override that method instead of patching the module name.

`scripts/check_layer_imports.py` is now clean (0 violations).
