+++ openrag/services/auth/__init__.py
@@ -1,9 +1,10 @@
 """Auth service layer - OIDC client, session tokens, state cookie, and deps."""
 
+from core.auth.state_cookie import StateCookiePayload, StateCookieSerializer
+
 from .deps import get_oidc_client, reset_oidc_client
 from .oidc_client import LogoutTokenClaims, OIDCClient, TokenBundle
 from .session_tokens import decrypt_token, encrypt_token, hash_session_token, issue_session_token
-from .state_cookie import StateCookiePayload, StateCookieSerializer
 
 __all__ = [
     "OIDCClient",
